requirements.txt
https://github.com/mraardvark/pyupdi/archive/master.zip
pySerial
